updating test_helpers
