+++ traitsui/testing/tester/qt4/tests/test_helpers.py
@@ -17,7 +17,10 @@ from traitsui.tests._tools import (
     requires_toolkit,
     ToolkitName,
 )
+from traitsui.testing.tester import command
+from traitsui.testing.tester.exceptions import Disabled
 from traitsui.testing.tester.qt4 import helpers
+
 
 try:
     from pyface.qt import QtGui
@@ -52,3 +55,52 @@ class TestInteractions(unittest.TestCase):
 
         # then
         self.assertEqual(click_slot.call_count, 0)
+
+    def test_key_sequence(self):
+        textbox = QtGui.QLineEdit()
+        change_slot = mock.Mock()
+        textbox.textEdited.connect(change_slot)
+
+        # when
+        helpers.key_sequence_qwidget(textbox, command.KeySequence("abc"), 0)
+
+        # then
+        self.assertEqual(textbox.text(), "abc")
+        # each keystroke fires a signal
+        self.assertEqual(change_slot.call_count, 3)
+
+
+    def test_key_sequence_disabled(self):
+        textbox = QtGui.QLineEdit()
+        textbox.setEnabled(False)
+        
+        # this will fail, because one should not be allowed to set
+        # cursor on the widget to type anything
+        with self.assertRaises(Disabled):
+            helpers.key_sequence_qwidget(textbox, command.KeySequence("abc"), 0)
+
+
+    def test_key_click(self):
+        textbox = QtGui.QLineEdit()
+        change_slot = mock.Mock()
+        textbox.editingFinished.connect(change_slot)
+        
+        # sanity check
+        helpers.key_sequence_qwidget(textbox, command.KeySequence("abc"), 0)
+        self.assertEqual(change_slot.call_count, 0)
+
+        helpers.key_click_qwidget(textbox, command.KeyClick("Enter"), 0)
+        self.assertEqual(change_slot.call_count, 1)
+
+
+    def test_key_click_disabled(self):
+        textbox = QtGui.QLineEdit()
+        textbox.setEnabled(False)
+        change_slot = mock.Mock()
+        textbox.editingFinished.connect(change_slot)
+
+
+        with self.assertRaises(Disabled):
+            helpers.key_click_qwidget(textbox, command.KeyClick("Enter"), 0)
+        self.assertEqual(change_slot.call_count, 0)
+
